backend/requirements.txt
fastapi
uvicorn
spacy
sumy
yake
nltk
pydantic
numpy
sentence-transformers
scikit-learn
torch
flask
flask-cors
requests
python-dotenv
pandas
tqdm
transformers
python-multipart
gunicorn